# openapi.yaml
openapi: "3.1.0"
info:
  title: "SENTRA API"
  version: "1.0.0"
paths:
  /files/read:
    post:
      tags:
        - "files"
      summary: "Files.Read"
      operationId: "files.read"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileReadRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileReadResponse"
        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /files/write:
    post:
      tags:
        - "files"
      summary: "Files.Write"
      operationId: "files.write"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileWriteRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileWriteResponse"
        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
  /n8n/trigger:
    post:
      tags:
        - "n8n"
      summary: "N8N.Trigger"
      operationId: "n8n.trigger"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/N8NTriggerRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/N8NTriggerResponse"
        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
  /bus/send:
    post:
      tags:
        - "bus"
      summary: "Bus.Send"
      operationId: "bus.send"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusSendRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusSendResponse"
        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /bus/poll:
    post:
      tags:
        - "bus"
      summary: "Bus.Poll"
      operationId: "bus.poll"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusPollRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusPollResponse"
        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /bus/updateStatus:
    post:
      tags:
        - "bus"
      summary: "Bus.Updatestatus"
      operationId: "bus.updateStatus"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusUpdateStatusRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusUpdateStatusResponse"
        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /rag/index:
    post:
      tags:
        - "rag"
      summary: "Rag.Index"
      operationId: "rag.index"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RAGIndexRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RAGIndexResponse"
        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /rag/query:
    post:
      tags:
        - "rag"
      summary: "Rag.Query"
      operationId: "rag.query"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RAGQueryRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RAGQueryResponse"
        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
components:
  schemas:
    BusPollRequest:
      properties:
        user:
          type: "string"
          minLength: 1
          title: "User"
        spreadsheet_id:
          type: "string"
          minLength: 1
          title: "Spreadsheet Id"
        worksheet:
          type: "string"
          minLength: 1
          title: "Worksheet"
        status:
          anyOf:
            -
              type: "string"
            -
              type: "null"
          title: "Status"
        limit:
          type: "integer"
          maximum: 100.0
          minimum: 1.0
          title: "Limit"
          default: 20
      type: "object"
      required:
        - "user"
        - "spreadsheet_id"
        - "worksheet"
      title: "BusPollRequest"
    BusPollResponse:
      properties:
        records:
          items:
            $ref: "#/components/schemas/BusRecord"
          type: "array"
          title: "Records"
      type: "object"
      required:
        - "records"
      title: "BusPollResponse"
    BusRecord:
      properties:
        message_id:
          type: "string"
          title: "Message Id"
        timestamp:
          type: "string"
          title: "Timestamp"
        user:
          type: "string"
          title: "User"
        agent:
          type: "string"
          title: "Agent"
        status:
          type: "string"
          title: "Status"
        payload:
          additionalProperties: true
          type: "object"
          title: "Payload"
      type: "object"
      required:
        - "message_id"
        - "timestamp"
        - "user"
        - "agent"
        - "status"
        - "payload"
      title: "BusRecord"
    BusSendRequest:
      properties:
        user:
          type: "string"
          minLength: 1
          title: "User"
        agent:
          type: "string"
          minLength: 1
          title: "Agent"
        spreadsheet_id:
          type: "string"
          minLength: 1
          title: "Spreadsheet Id"
        worksheet:
          type: "string"
          minLength: 1
          title: "Worksheet"
        payload:
          additionalProperties: true
          type: "object"
          title: "Payload"
        idempotency_key:
          anyOf:
            -
              type: "string"
            -
              type: "null"
          title: "Idempotency Key"
      type: "object"
      required:
        - "user"
        - "agent"
        - "spreadsheet_id"
        - "worksheet"
      title: "BusSendRequest"
    BusSendResponse:
      properties:
        message_id:
          type: "string"
          title: "Message Id"
        status:
          type: "string"
          title: "Status"
        timestamp:
          type: "string"
          title: "Timestamp"
      type: "object"
      required:
        - "message_id"
        - "status"
        - "timestamp"
      title: "BusSendResponse"
    BusUpdateStatusRequest:
      properties:
        user:
          type: "string"
          minLength: 1
          title: "User"
        agent:
          type: "string"
          minLength: 1
          title: "Agent"
        spreadsheet_id:
          type: "string"
          minLength: 1
          title: "Spreadsheet Id"
        worksheet:
          type: "string"
          minLength: 1
          title: "Worksheet"
        message_id:
          type: "string"
          minLength: 1
          title: "Message Id"
        status:
          type: "string"
          minLength: 1
          title: "Status"
      type: "object"
      required:
        - "user"
        - "agent"
        - "spreadsheet_id"
        - "worksheet"
        - "message_id"
        - "status"
      title: "BusUpdateStatusRequest"
    BusUpdateStatusResponse:
      properties:
        message_id:
          type: "string"
          title: "Message Id"
        status:
          type: "string"
          title: "Status"
      type: "object"
      required:
        - "message_id"
        - "status"
      title: "BusUpdateStatusResponse"
    FileReadRequest:
      properties:
        path:
          type: "string"
          title: "Path"
          description: "Path beginning with /projects, /reports, or /students"
        user:
          type: "string"
          minLength: 1
          title: "User"
      type: "object"
      required:
        - "path"
        - "user"
      title: "FileReadRequest"
    FileReadResponse:
      properties:
        path:
          type: "string"
          title: "Path"
        content:
          type: "string"
          title: "Content"
        sha256:
          type: "string"
          title: "Sha256"
        last_modified:
          anyOf:
            -
              type: "string"
              format: "date-time"
            -
              type: "null"
          title: "Last Modified"
      type: "object"
      required:
        - "path"
        - "content"
        - "sha256"
        - "last_modified"
      title: "FileReadResponse"
    FileWriteRequest:
      properties:
        path:
          type: "string"
          title: "Path"
          description: "Path beginning with /projects, /reports, or /students"
        user:
          type: "string"
          minLength: 1
          title: "User"
        content:
          type: "string"
          title: "Content"
          description: "File contents to persist"
        agent:
          type: "string"
          minLength: 1
          title: "Agent"
        idempotency_key:
          anyOf:
            -
              type: "string"
            -
              type: "null"
          title: "Idempotency Key"
          description: "Caller-supplied idempotency key"
      type: "object"
      required:
        - "path"
        - "user"
        - "content"
        - "agent"
      title: "FileWriteRequest"
    FileWriteResponse:
      properties:
        path:
          type: "string"
          title: "Path"
        sha256:
          type: "string"
          title: "Sha256"
        committed:
          type: "boolean"
          title: "Committed"
        commit_message:
          anyOf:
            -
              type: "string"
            -
              type: "null"
          title: "Commit Message"
      type: "object"
      required:
        - "path"
        - "sha256"
        - "committed"
        - "commit_message"
      title: "FileWriteResponse"
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          type: "array"
          title: "Detail"
      type: "object"
      title: "HTTPValidationError"
...
    N8NTriggerRequest:
      properties:
        user:
          type: "string"
          minLength: 1
          title: "User"
        agent:
          type: "string"
          minLength: 1
          title: "Agent"
        payload:
          additionalProperties: true
          type: "object"
          title: "Payload"
        idempotency_key:
          anyOf:
            -
              type: "string"
            -
              type: "null"
          title: "Idempotency Key"
      type: "object"
      required:
        - "user"
        - "agent"
      title: "N8NTriggerRequest"
    N8NTriggerResponse:
      properties:
        result:
          additionalProperties: true
          type: "object"
          title: "Result"
      type: "object"
      required:
        - "result"
      title: "N8NTriggerResponse"
    RAGDocumentPayload:
      properties:
        text:
          type: "string"
          minLength: 1
          title: "Text"
        metadata:
          additionalProperties: true
          type: "object"
          title: "Metadata"
        id:
          anyOf:
            -
              type: "string"
            -
              type: "null"
          title: "Id"
      type: "object"
      required:
        - "text"
      title: "RAGDocumentPayload"
    RAGIndexRequest:
      properties:
        user:
          type: "string"
          minLength: 1
          title: "User"
        agent:
          type: "string"
          minLength: 1
          title: "Agent"
        collection:
          type: "string"
          minLength: 1
          title: "Collection"
        documents:
          items:
            $ref: "#/components/schemas/RAGDocumentPayload"
          type: "array"
          title: "Documents"
      type: "object"
      required:
        - "user"
        - "agent"
        - "collection"
        - "documents"
      title: "RAGIndexRequest"
    RAGIndexResponse:
      properties:
        document_ids:
          items:
            type: "string"
          type: "array"
          title: "Document Ids"
      type: "object"
      required:
        - "document_ids"
      title: "RAGIndexResponse"
    RAGQueryRequest:
      properties:
        user:
          type: "string"
          minLength: 1
          title: "User"
        collection:
          type: "string"
          minLength: 1
          title: "Collection"
        query:
          type: "string"
          minLength: 1
          title: "Query"
        n_results:
          type: "integer"
          maximum: 50.0
          minimum: 1.0
          title: "N Results"
          default: 5
      type: "object"
      required:
        - "user"
        - "collection"
        - "query"
      title: "RAGQueryRequest"
    RAGQueryResponse:
      properties:
        results:
          additionalProperties: true
          type: "object"
          title: "Results"
      type: "object"
      required:
        - "results"
      title: "RAGQueryResponse"
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              -
                type: "string"
              -
                type: "integer"
          type: "array"
          title: "Location"
        msg:
          type: "string"
          title: "Message"
        type:
          type: "string"
          title: "Error Type"
      type: "object"
      required:
        - "loc"
        - "msg"
        - "type"
      title: "ValidationError"
servers:
  -
    url: "http://localhost:5000"